TFP-31 Extend TagMappingTrait with data Provider and reuse it
