+++ tests/src/Kernel/TagMappingTest.php
@@ -11,6 +11,8 @@ use Drupal\KernelTests\KernelTestBase;
  */
 class TagMappingTest extends KernelTestBase {
 
+  use TagMappingTrait;
+
   /**
    * Modules to enable.
    *
@@ -35,61 +37,6 @@ class TagMappingTest extends KernelTestBase {
     $tag_xmltag_story = $storage->create($data);
     $tag_xmltag_story->validate();
     $tag_xmltag_story->save();
-  }
-
-  /**
-   * Data provider for tag mapping config data.
-   *
-   * @return array
-   */
-  public function tagMappingProvider() {
-    return [
-      'title' => [
-        [
-          'id' => 'xmltag_story',
-          'mapping_type' => 'text_plain',
-          'mapping' => [
-            [
-              'property' => 'value',
-              'tag' => 'XMLTag/Title',
-            ],
-          ],
-          'options' => [
-            'title' => TRUE,
-          ],
-        ]
-      ],
-      'long_text' => [
-        [
-          'id' => 'xmltag_story',
-          'mapping_type' => 'text_formatted_long',
-          'mapping' => [
-            [
-              'property' => 'value',
-              'tag' => 'XMLTag/Story',
-            ],
-          ],
-          'options' => [],
-        ]
-      ],
-      'image' => [
-        [
-          'id' => 'xmltag_image',
-          'mapping_type' => 'media_image',
-          'mapping' => [
-            [
-              'property' => 'field_image',
-              'tag' => 'XMLTag/Image',
-            ],
-            [
-              'property' => 'field_description',
-              'tag' => 'XMLTag/Caption',
-            ],
-          ],
-          'options' => [],
-        ]
-      ]
-    ];
   }
